% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brush.R
\name{brushed}
\alias{brushed}
\title{Returns a data frame with points marked which are under a brush}
\usage{
brushed(plotExpr, allRows = FALSE)
}
\arguments{
\item{plotExpr}{A ggplot object.}

\item{allRows}{If FALSE (the default) return a data frame containing the
selected rows. If TRUE, the input data frame will have a new column,
selected_, which indicates whether the row was inside the brush (TRUE) or
outside the brush (FALSE).}
}
\value{
A data frame
}
\description{
This function returns all rows from the input data frame with an additional
column selected_, which indicates which rows of the input data frame are
selected by the brush (TRUE for selected, FALSE for not-selected).The xvar,
yvar arguments specify which columns in the data correspond to the x
variable, y variable of the plot. Based on ggbrush by Joe Cheng.
}
\examples{
library(lubridate)
library(ggplot2)

# Example dataset
dates = as.POSIXct(seq(from = ymd("2013-01-01"),
 to = ymd("2014-01-01"), by = 3 * 3600))
Data = data.frame(dates, value = sin(decimal_date(dates)/0.01) +
 rnorm(length(dates)))

# Select points
p = qplot(data = Data, dates, value)
brushed(p)

}
