% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editData.R
\name{editData}
\alias{editData}
\title{Edit and flag data}
\usage{
editData(df, Subset = 1:nrow(Data), col.name = "DataValue", flag = "",
  FUN = function(x) x, ...)
}
\arguments{
\item{df}{dataframe}

\item{Subset}{subset    of values to apply edit}

\item{col.name}{name of the column to edit}

\item{flag}{flag to apply to edited data record}

\item{FUN}{the function to be applied}
}
\description{
The function editData can be used to edit and flag data. A subset of values
can be targeted and worked on.
}
\details{
The function supplied (FUN) will be applied across the values defined by the
date range given to the Subset argument. The full data set will be returned
with the edits applied and values flagged.
}
\examples{

library(dplyr)
library(lubridate)

# Example dataset
dates = as.POSIXct(seq(from = ymd("2013-01-01"),
 to = ymd("2014-01-01"), by = 3 * 3600))
Data = data.frame(dates, value = sin(decimal_date(dates)/0.01) +
 rnorm(length(dates)))

# set flag
Data = editData(Data, col.name = "value", flag = "missing")
# set a constant value
Data = editData(Data, col.name = "value", FUN = function (x) 5)
# apply offset
Data = editData(Data, col.name = "value", FUN = function (x) x + 5)
# subset data
Data = editData(Data, col.name = "value", between(Data$dates,
 ymd("2013-01-01"), ymd("2013-02-02")),
 FUN = function (x) x + -5,
 flag = "offset -5")

}

